Rename time module in amd64 to tsc
